support CPU Adam and Adagrad on Windows with SDK 10.0.22000
