+++ csrc/includes/cpu_adagrad.h
@@ -1,5 +1,8 @@
 #pragma once
 
+#define NOMINMAX  // Windows idiosyncrasy
+                  // https://stackoverflow.com/questions/4913922/possible-problems-with-nominmax-on-visual-c
+
 #include <cuda_fp16.h>
 #include <cuda_runtime_api.h>
 #include <stdio.h>
